move controller to a separate mod
